removed the incremental flag
change the annoying throws Exception to throws IOException

removed null pointer issue in the calculation of the xml base.

put a try finally to ensurre that event generation is restored even if
things go wrong.

ensured that an ontology is generated if it is not in the parse.

added sortSubclasses and copyFacetValuesIntoNamedClses postprocessing.