Fix for all InstanceGridWidget class and subclasses (including the linearization, postcoordination, scale value, inclusion and index widgets) that would sometimes not work properly because the store and shadow store were not in sync. This was evident in the Index widget when the text contained uppercase letters (because of the sorting in the initial solution, and the way that sorting is done in the stores: everything is uppercased first).
The fix is to create an additional field in the store that links to the correspondent record in the shadow store. In this way, we are always sure that the store and the shadow stores are "in sync". The stores/widgets can now also be sorted.

This is a major fix with implications for many important widgets.