1 major change and 2 minor changes:
(1) Some of the widgets in the PropertyFieldPortlet are now filled with a bulk remote call handled by the GetEntityTriplesHandler and the GetEntityPropertyValuesHandler. This way, we reduce a lot the number of remote calls made. This is a major change. The change should be backwards compatible with widgets that do not support the bulk remote calls.

(2) Started experimenting with the ResourceBundles, and using the in the InstanceGridWidget for the loading status. But, we might expand them in the future.

(3) In OntologyServiceImpl: 
- the search methods now return entity lists with types, so that the UI can update correctly when jumping.
- added a few more util methods for creating EntityData and lists thereof.